requirements.txt
gym==0.21.0
imageio==2.37.0
matplotlib==3.10.0
numpy==1.25.2
ollama==0.4.7
opensimplex==0.4.5.1
Pillow==11.1.0
pygame==2.6.1
ruamel.base==1.0.0
stable_baselines3==2.5.0
openai==1.95.0
opencv-python
